+++ amplpy/ampl.pyx
@@ -10,6 +10,7 @@ from cpython.bool cimport PyBool_Check
 
 
 from numbers import Real
+from ast import literal_eval
 
 include "util.pxi" # must be first
 include "constraint.pxi"
@@ -24,6 +25,22 @@ include "outputhandler.pxi"
 include "parameter.pxi"
 include "set.pxi"
 include "variable.pxi"
+
+
+def nested_dict_of_suffixes(lst):
+    nested = {}
+    for name, value in lst:
+        if "[" not in name:
+            nested[name] = value
+        else:
+            p = name.find("[")
+            v, index = name[:p], literal_eval(f"({name[p+1:-1]},)")
+            if v not in nested:
+                nested[v] = {}
+            if len(index) == 1:
+                index = index[0]
+            nested[v][index] = value
+    return nested
 
 
 AMPL_NOT_FOUND_MESSAGE = """
@@ -915,6 +932,79 @@ cdef class AMPL:
         +---------+-----------+---------------------------------------------+
         """
         return self.get_value("solve_result_num")
+
+    def get_iis(self, flat=True):
+        """
+        Get IIS attributes for all variables and constraints.
+
+        Args:
+            flat: Return flat dictionaries if set to True, or nested dictionaries otherwise.
+
+        Returns:
+            Tuple with a dictionary for variables in the IIS and another for the constraints.
+
+        Usage example:
+
+        .. code-block:: python
+
+            from amplpy import AMPL
+            ampl = AMPL()
+            ampl.eval(
+                r\"\"\"
+            var x >= 0;
+            var y >= 0;
+            maximize obj: x+y;
+            s.t. s: x+y <= -5;
+            \"\"\"
+            )
+            ampl.option["presolve"] = 0  # disable AMPL presolve
+            ampl.solve(solver="gurobi", gurobi_options="outlev=1 iis=1")
+            if ampl.solve_result == "infeasible":
+                var_iis, con_iis = ampl.get_iis()
+                print(var_iis, con_iis)
+        """
+        iis_var = self.get_data(
+            "{i in 1.._nvars: _var[i].iis != 'non'} (_varname[i], _var[i].iis)"
+        ).to_list(skip_index=True)
+        iis_con = self.get_data(
+            "{i in 1.._ncons: _con[i].iis != 'non'} (_conname[i], _con[i].iis)"
+        ).to_list(skip_index=True)
+        if flat is False:
+            iis_var = nested_dict_of_suffixes(iis_var)
+            iis_con = nested_dict_of_suffixes(iis_con)
+        else:
+            iis_var = dict(iis_var)
+            iis_con = dict(iis_con)
+        return iis_var, iis_con
+
+    def get_solution(self, flat=True, zeros=False):
+        """
+        Get solution values for all variables.
+
+        Args:
+            flat: Return a flat dictionary if set to True, or a nested dictionary otherwise.
+
+            zeros: Include zeros in the solution if set to True.
+
+        Returns:
+            Returns a dictionary with the solution.
+
+        Usage example:
+
+        .. code-block:: python
+
+            ampl.solve(solver="gurobi", gurobi_options="outlev=0")
+            if ampl.solve_result == "solved":
+                print(ampl.get_solution())
+        """
+        if zeros:
+            stmt = "{i in 1.._nvars} (_varname[i], _var[i].val)"
+        else:
+            stmt = "{i in 1.._nvars: _var[i].val != 0} (_varname[i], _var[i].val)"
+        lst_solution = self.get_data(stmt).to_list(skip_index=True)
+        if flat:
+            return dict(lst_solution)
+        return nested_dict_of_suffixes(lst_solution)
 
     def _start_recording(self, filename):
         """